Add useIsValid hook for draft validation and remove legacy isValid utility
- Introduced the `useIsValid` hook to streamline draft validation by leveraging the `ListingCreateSchema` for improved data integrity.
- Updated components to utilize the new `useIsValid` hook, enhancing consistency in draft validation across the application.
- Removed the legacy `isValid` utility function to reduce redundancy and improve code maintainability.
